requirements.txt
fastapi==0.104.1
fastapi-poe==0.0.36
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
pinecone==5.0.0
pypdf2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.1
httpx==0.25.0